Fixed z score normalization
The z score normalization was applied to each *feature* -- in this
context the lagged Swiss milk price -- individually.

However, when re-scaling the predicted values, the same mean $\mu$ and
standard deviation $\sigma$ were used for all forecast horizons.

This lead to slight misscalculations in the forecast; it was fixed by
using the same mean and standard deviations for each forecast horizon.